fix: Profile view tests
